DEV(cache_critical_dns): add caching for MessageBus Redis hostname
We are already caching any DB_HOST and REDIS_HOST (and their
accompanying replicas), we should also cache the resolved addresses for
the MessageBus specific Redis. This is a noop if no MB redis is defined
in config.

A side effect is that the MB will also support SRV lookup and
priorities, following the same convention as the other cached services.

See /t/68301/30.